refactor(engine): extract wwebjs lifecycle collaborators from the adapter
Split the whatsapp-web.js adapter along its remaining cohesive seams, following
the delegate conventions the directory already uses (WwebjsEngineHost slices
for the API domains, BaileysLifecycle for connection state on the Baileys
side). Four new collaborators plus two extended ones:

- wwebjs-lifecycle: initialize (with the navigation-killed single retry),
  client-event wiring for connection state, puppeteer death detection, the
  four teardown paths, the liveness probe, the pairing-code gate, and the
  connection state itself.
- wwebjs-reconcile: the post-authentication readiness reconciliation window,
  its runtime probe, and the one-shot bridge reload.
- wwebjs-stuck-auth: the stuck-auth claim/recovery path and the LocalAuth
  profile removal both it and a WhatsApp-initiated logout share.
- wwebjs-calls: the incoming-call event, the live-call cache, rejectCall.
- wwebjs-onboarding: gains the watcher loop around its existing in-page probe.
- wwebjs-group-events: gains the group-notification event registration around
  its existing payload mappers.

The adapter keeps the full IWhatsAppEngine surface as thin forwarders and
aliases the spec-pinned state fields by reference, so behavior is unchanged.
Exported helpers and constants (isExecutionContextDestroyedError, the
reconcile/navigation timeouts, proxy/messaging/onboarding helpers) are
re-exported from the adapter module for existing callers.
